requirements.txt
tushare>=1.2.80
python-telegram-bot>=20.0
requests>=2.28.0
python-dotenv>=0.19.0
pandas>=1.3.0
